+++ src/chart/mod.rs
@@ -18,5 +18,6 @@ mod mesh;
 mod series;
 
 pub use builder::ChartBuilder;
-pub use context::ChartContext;
+pub use context::{ChartContext, SeriesAnno};
 pub use mesh::MeshStyle;
+pub use series::SeriesLabelStyle;