+++ CHANGELOG.md
@@ -6,10 +6,12 @@
 - Font is now support rotation transformation. Use `FontDesc::transform` to apply an rotation to transformation. For example, `font.transform(FontTransform::Rotate90)`.
 - ChartContext now support drawing axis description. Use `MeshStyle::x_desc` and `MeshStyle::y_desc` to specify the axis description text.
 - Add series label support. `ChartContext::draw_series` now returns a struct `SeriesAnno` that collects the additional information for series labeling. `ChartContext::draw_series_labels` are used to actually draw the series label. (See `examples/chart.rs` for detailed examples)
+- Mocking drawing backend.
 
 ### Improvement
 - Unify `OwnedText` and `Text` into `Text`. Previously, `OwnedText` and `Text` are two seperate types, one holds a `String` another holds a `&str`. Now `OwnedText` is removed.
 use `Text::new("text".to_string(),...)` for owned text element and `Text::new("text", ...)` for borrowed text.
+- More test cases
 
 ## Plotters 0.1.13 (2019-05-31)